update so taht we are starting fresh
